python/requirements.txt
-i https://pypi.org/simple
absl-py==1.3.0; python_version >= '3.6'
astunparse==1.6.3
cachetools==5.2.0; python_version ~= '3.7'
certifi==2022.12.7; python_version >= '3.6'
charset-normalizer==2.1.1; python_version >= '3.6'
flatbuffers==22.12.6
gast==0.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-auth==2.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-auth-oauthlib==0.4.6; python_version >= '3.6'
google-pasta==0.2.0
grpcio==1.51.1; python_version >= '3.7'
h5py==3.7.0; python_version >= '3.7'
idna==3.4; python_version >= '3.5'
imageio==2.23.0; python_version >= '3.7'
keras==2.11.0; python_version >= '3.7'
libclang==14.0.6
markdown==3.4.1; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
networkx==2.8.8; python_version >= '3.8'
numpy==1.24.1
numpy-quaternion==2022.4.2
oauthlib==3.2.2; python_version >= '3.6'
opencv-python==4.6.0.66
opt-einsum==3.3.0; python_version >= '3.5'
packaging==22.0; python_version >= '3.7'
pillow==9.3.0; python_version >= '3.7'
protobuf==3.19.6; python_version >= '3.5'
pyasn1==0.4.8
pyasn1-modules==0.2.8
pywavelets==1.4.1; python_version >= '3.8'
requests==2.28.1; python_version >= '3.7' and python_version < '4'
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rsa==4.9; python_version >= '3.6'
scikit-image==0.19.3
scipy==1.9.3; python_version >= '3.8'
setuptools==65.6.3; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tensorboard==2.11.0; python_version >= '3.7'
tensorboard-data-server==0.6.1; python_version >= '3.6'
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.19.0
tensorflow-estimator==2.11.0; python_version >= '3.7'
tensorflow-gpu==2.11.0
tensorflow-io-gcs-filesystem==0.29.0; platform_machine != 'arm64' or platform_system != 'Darwin'
termcolor==2.1.1; python_version >= '3.7'
tifffile==2022.10.10; python_version >= '3.8'
typeguard==2.13.3; python_full_version >= '3.5.3'
typing-extensions==4.4.0; python_version >= '3.7'
urllib3==1.26.13; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
werkzeug==2.2.2; python_version >= '3.7'
wheel==0.38.4; python_version >= '3.7'
wrapt==1.14.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
